I made a mess of shitty code
Made an attempt at login page result and failed. Requirement state that we should have a login status on home page. Will review this with group tomorrow. I think registration page is a more logical start based on the requirements and what we've learned so far so I'm going to work on the registration page before the meeting in the tomorrow.